[nrf noup] boards: add nrf54l15/nrf54l15/cpuapp/ns as board variant
This commit adds the nRF54l15//nrf54l15/cpuapp/ns board variant to
sdk-zephyr. It allows to build nRF54l15pdk with TF-M in nRF Connect SDK.

This is done as a nrf noup patch as the out-of-tree board extension
feature is not available. When zephyrproject-rtos/zephyr#69548 is
completed, then this commit can be reverted and the variant be placed in
sdk-nrf repository.

The nRF54L15 is using the SAU and the MPC to configure the security
attributes for TrustZone. The SPU is only used for the configuration
of the peripherals. This sets the Kconfig configurations for TrustZone
which use the correct region size(based on the MPC) for nRF54L15.

Ref: NCSDK-25023

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>